Improve http verification script
As we got two copies of same script, we need to sync them at some point.
If we have same script, it can be removed from openQA distri repo. Our
aim is to have only a single copy of profiles and verification scripts.
This version of script is more readable and matches http profile we use
in openQA, which is also sligthly different to the one in aytests
package. TBD how to proceed.