Move or sync profiles and verification scripts
As part of improvement of autoyast tests, we want to keep profiles and
verification scripts in single place. As same profiles are used for
integration tests, we move them to aytests repository.
Whereas, as some files were forked, we need to sync them, to be on the
same side. After merging these changes, settings of relevant jobs have
to be changed. There will be another merge required for
os-autoinst-distri-opensuse repository.